- Moved ontology-configuration into GUI and all the imports are now relative - Adopted gui.py (main section) to supply the current ontology and that ontology is present in line ~70 of ontology_configuration_extended and it gets translated into v3 - When I click on load it is shown
Hints for fast testing:
- start pasta usual way:
- restart with F9
- open ontology config F8 (for now and for testing)
