+++ pasta_eln/GUI/ontology_configuration/__init__.py
@@ -6,3 +6,10 @@
 #  Filename: __init__.py
 #
 #  You should have received a copy of the license with this file. Please refer the license file for more information.
+# Ontology Table view constants
+PROPS_TABLE_REQUIRED_COLUMN_INDEX = 4
+PROPS_TABLE_DELETE_COLUMN_INDEX = 6
+PROPS_TABLE_REORDER_COLUMN_INDEX = 7
+
+ATTACHMENT_TABLE_DELETE_COLUMN_INDEX = 2
+ATTACHMENT_TABLE_REORDER_COLUMN_INDEX = 3
