+++ pasta_eln/GUI/ontology_configuration/create_type_dialog/create_type_dialog.py
