+++ pasta_eln/GUI/ontology_configuration/delete_column_delegate.py
@@ -1,28 +1,29 @@
-""" ReorderColumnDelegate for the table views """
+""" DeleteColumnDelegate for the table views """
 #  PASTA-ELN and all its sub-parts are covered by the MIT license.
 #
 #  Copyright (c) 2023
 #
 #  Author: Jithu Murugan
-#  Filename: reorder_column_delegate.py
+#  Filename: delete_column_delegate.py
 #
 #  You should have received a copy of the license with this file. Please refer the license file for more information.
 import logging
 from typing import Union
 
-from PySide6.QtCore import QModelIndex, QPersistentModelIndex, Signal, QEvent, QAbstractItemModel
+from PySide6.QtCore import QModelIndex, QPersistentModelIndex, QEvent, QAbstractItemModel, Signal
 from PySide6.QtGui import QPainter
 from PySide6.QtWidgets import QStyledItemDelegate, QPushButton, QWidget, QStyleOptionViewItem, QStyleOptionButton, \
   QStyle, QApplication
 
-from pasta_eln.ontology_configuration.utility_functions import is_click_within_bounds
+from .utility_functions import is_click_within_bounds
 
 
-class ReorderColumnDelegate(QStyledItemDelegate):
+class DeleteColumnDelegate(QStyledItemDelegate):
   """
-  Delegate for creating the icons for the re-order column in the ontology editor tables
+  Delegate for creating the delete icon for the delete column in the ontology table views
   """
-  re_order_signal = Signal(int)
+  delete_clicked_signal = Signal(
+    int)  # Signal to inform the delete button click with the position in the table as the parameter
 
   def __init__(self) -> None:
     """
@@ -36,11 +37,11 @@ class ReorderColumnDelegate(QStyledItemDelegate):
             option: QStyleOptionViewItem,
             index: Union[QModelIndex, QPersistentModelIndex]) -> None:
     """
-    Draws the required re-order button within the cell represented by index
+    Draws the delete button within the cell represented by index
     Args:
       painter (QPainter): Painter instance for painting the button.
       option (QStyleOptionViewItem): Style option for the cell represented by index.
-      index (Union[QModelIndex, QPersistentModelIndex]): Table cell index
+      index (Union[QModelIndex, QPersistentModelIndex]): Cell index.
 
     Returns: None
 
@@ -49,7 +50,7 @@ class ReorderColumnDelegate(QStyledItemDelegate):
     opt = QStyleOptionButton()
     opt.state = QStyle.State_Active | QStyle.State_Enabled
     opt.rect = option.rect
-    opt.text = "^"
+    opt.text = "Delete"
     QApplication.style().drawControl(QStyle.CE_PushButton, opt, painter, button)
 
   def createEditor(self,
@@ -57,7 +58,7 @@ class ReorderColumnDelegate(QStyledItemDelegate):
                    option: QStyleOptionViewItem,
                    index: Union[QModelIndex, QPersistentModelIndex]) -> None:
     """
-    Disable the editor for the whole re-order column by simply returning None
+    Disable the editor for the delete column by simply returning None
     Args:
       parent (QWidget): Parent table view.
       option (QStyleOptionViewItem): Style option for the cell represented by index.
@@ -74,19 +75,19 @@ class ReorderColumnDelegate(QStyledItemDelegate):
                   option: QStyleOptionViewItem,
                   index: Union[QModelIndex, QPersistentModelIndex]) -> bool:
     """
-    In case of mouse click event, the re_order_signal is emitted for the respective table cell position
+    In case of click detected within the cell represented by index, the respective delete signal is emitted
     Args:
       event (QEvent): The editor event information.
       model (QAbstractItemModel): Model data representing the table view.
       option (QStyleOptionViewItem): QStyleOption for the table cell.
       index (Union[QModelIndex, QPersistentModelIndex]): Table cell index.
 
-    Returns (bool): True/False
+    Returns (bool): True if deleted otherwise False
 
     """
     if is_click_within_bounds(event, option):
       row = index.row()
-      self.logger.info("Re-order signal emitted for the position: {%s} in the table..", row)
-      self.re_order_signal.emit(row)
+      self.logger.info("Delete signal emitted for the position: {%s}", row)
+      self.delete_clicked_signal.emit(row)
       return True
     return False
