+++ pasta_eln/GUI/ontology_configuration/exceptions/ontology_config_generic_exception.py
