+++ pasta_eln/GUI/ontology_configuration/utility_functions.py
